Initial project setup
Created an initial project setup for a C++ project. The build is managed
by CMake, package management is done with vcpkg, code linting and
formatting is implemented by clang-tidy and code testing with GTest.

CMake was chosen due to its great portability to different operating
systems to make the build as portable as it can get.

Also, vcpkg was chosen as an dependency management tool as it gained a
lot of traction in the C++ community for its ease of use.

To ensure code quality, clang-tidy is used to check code formatting,
linting and also make some static analysis.

Furthermore, GTest is used to enable testing to enhance code stability
and making sure that all code pieces get tested.

Currently, this is only a project skeleton. Implementation will follow.